split up chatPicker file
