+++ emphasis_education_app/src/components/Chat/ChatPicker.tsx
@@ -1,15 +1,12 @@
 import * as React from 'react';
 import {
-  Alert,
   Text,
   View,
   SafeAreaView,
-  TouchableOpacity,
   ActivityIndicator,
 } from 'react-native';
 import { Icon } from 'react-native-elements';
-import styled from 'styled-components';
-import { useLazyQuery, useMutation } from '@apollo/react-hooks';
+import { useLazyQuery } from '@apollo/react-hooks';
 import messaging from '@react-native-firebase/messaging';
 
 import { GeneralContext } from '../Context/Context';
@@ -17,278 +14,23 @@ import {
   PermissionedComponent,
   IconRow,
   GeneralSpacing,
-  ThemedText,
-  VerticalDivider,
-  HorizontalDivider,
-  FONT_STYLES,
 } from '../shared';
 import { GET_USER } from '../../queries/GetUser';
-import { DELETE_CHAT } from '../../queries/DeleteChat';
-import {
-  Permission,
-  Class,
-  ChatUserInfo,
-  IUser
-} from '../../types';
-import { theme } from '../../theme';
+import { Permission, ChatUserInfo, } from '../../types';
 
 import { EmptyChatPicker } from './common';
+import IndividualChat from './IndividualChat';
 
 interface IChatPickerProps {
   navigation: any;
   route: any;
 }
 
-const IndChat = styled(TouchableOpacity)`
-  border: black;
-  width: 100%;
-  height: 30px;
-`;
-
 export const ChatsContain: React.FC = ({ children }) => (
   <GeneralSpacing u={0} r={10} d={0} l={10}>
     {children}
   </GeneralSpacing>
 );
-
-
-const IconRowLeft = styled(IconRow)`
-  justify-content: flex-start;
-`;
-
-const LeftText = styled(ThemedText)`
-  padding-right: 5px;
-`
-
-const RightText = styled(ThemedText)`
-  padding-left: 5px;
-`
-
-// const Contain = styled(View)`
-//   padding: 10px;
-//   width: 80%;
-//   border-bottom-width: 1px;
-//   border-bottom-color: black;
-// `
-
-const SpacedItemRow = styled(View)`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-`;
-
-const ChatContain: React.FC = ({ children }) => (
-  <GeneralSpacing u={10} r={10} d={10} l={10}>
-    {children}
-  </GeneralSpacing>
-);
-interface IChatDisplay {
-  className: string;
-  chatID: string
-  mainText: string;
-  secondaryText?: string;
-  caption?: string;
-  tutorInfo: ChatUserInfo;
-  userInfo: ChatUserInfo[];
-  displayNotificationBadge: boolean;
-  goToChat (sub: string, sec: string, tutorInfo: ChatUserInfo, userInfo: ChatUserInfo[]): void;
-  getClasses (): void;
-  clearNotificationCounter (chatID: string): void;
-}
-
-const LoadingScreen: React.FC<{ loading: boolean }> = ({ loading }) => (
-  <View>
-    <Text>
-      the chat is deleting
-    </Text>
-    <ActivityIndicator animating={loading} />
-  </View>
-)
-
-const Badge = styled(View)`
-  height: 20px;
-  width: 20px;
-  border-radius: 10px;
-  background-color: grey;
-  margin-right: 10px;
-`
-
-const NotificationBadge: React.FC = () => (
-  <Badge />
-)
-
-const ChatDisplay: React.FC<IChatDisplay> = ({
-  mainText,
-  secondaryText,
-  caption,
-  chatID,
-  goToChat,
-  className,
-  getClasses,
-  tutorInfo,
-  userInfo,
-  displayNotificationBadge,
-  clearNotificationCounter
- }) => {
-
-  const [delChat, { data, loading, error }] = useMutation(DELETE_CHAT, {
-    onCompleted: () => {
-      // console.log('done deleting chat')
-      // getClasses();
-    }
-  });
-
-  const options = { variables: { chatID } };
-
-  const triggerMutation = () => delChat(options)
-
-  const triggerDeleteAlert = () =>  (
-    Alert.alert(
-      'Confirm Delete',
-      'Are you sure you want to delete this chat?',
-      [
-        {
-          text: 'Delete',
-          onPress: () => triggerMutation()
-        },
-        {
-          text: 'Cancel',
-          onPress: () => console.log('canceled'),
-          style: 'cancel'
-        },
-      ]
-    )
-  )
-
-  const onPress = (
-    chatID: string,
-    className: string,
-    tutorInfo: ChatUserInfo,
-    userInfo: ChatUserInfo[]
-  ) => () => {
-    clearNotificationCounter(chatID);
-    goToChat(chatID, className, tutorInfo, userInfo)
-  }
-
-  return (
-    <>
-      { loading ? <LoadingScreen loading={loading} /> : (
-        <ChatContain>
-          <SpacedItemRow>
-            <IconRow>
-              {displayNotificationBadge && <NotificationBadge />}
-              <TouchableOpacity onPress={onPress(chatID, className, tutorInfo, userInfo)} onLongPress={() => console.log('long press')}>
-                <IconRowLeft>
-                  <LeftText size={18} type={FONT_STYLES.MAIN}>
-                    {mainText}
-                  </LeftText>
-                  {
-                    !!secondaryText && (
-                      <>
-                        <VerticalDivider height={15} />
-                        <RightText size={18} type={FONT_STYLES.MAIN}>
-                          {secondaryText}
-                        </RightText>
-                      </>
-                    )
-                  }
-                </IconRowLeft>
-                {
-                  caption && (
-                    <ThemedText size={14} type={FONT_STYLES.LIGHT}>
-                      {caption}
-                    </ThemedText>
-                  )
-                }
-              </TouchableOpacity>
-            </IconRow>
-
-            <PermissionedComponent
-              allowedPermissions={[Permission.Admin]}
-            >
-              <Icon
-                name='trash-o'
-                type='font-awesome'
-                onPress={triggerDeleteAlert}
-              />
-            </PermissionedComponent>
-          </SpacedItemRow>
-          <HorizontalDivider width={100} color={theme.colors.lightOrange}/>
-        </ChatContain>
-      )}
-    </>
-  )
-}
-
-interface IIndividualChatProps {
-  chatID: string;
-  userType: Permission;
-  classObject: Class;
-  displayNotificationBadge: boolean;
-  goToChat (sub: string, sec: string, tutorInfo: ChatUserInfo, userInfo: ChatUserInfo[]): void;
-  getClasses (): void;
-  clearNotificationCounter (chatID: string): void;
-}
-
-const IndividualChat: React.FC<IIndividualChatProps> = ({
-  classObject,
-  userType,
-  chatID,
-  goToChat,
-  getClasses,
-  displayNotificationBadge,
-  clearNotificationCounter
-}) => {
-
-  let caption;
-  let mainText: string = '';
-  let secondaryText: string = '';
-  // console.log('userType', userType)
-  // console.log('classObject', classObject)
-  const userFirstName: string[] = classObject.userInfo.map(_user => _user.firstName);
-
-  switch(userType) {
-    case 'Student':
-      mainText = classObject.className;
-      secondaryText = classObject.tutorInfo.firstName
-      caption = 'test caption'
-      break;
-    case 'Parent':
-      mainText = `${userFirstName.join(', ')}`
-      secondaryText = classObject.className
-      caption = `${classObject.tutorInfo.firstName} ${classObject.tutorInfo.lastName}`
-
-      break;
-    case 'Tutor':
-      mainText = `${userFirstName.join(', ')}`
-      caption = classObject.className
-      break
-    case 'Admin':
-      mainText = `${userFirstName.join(', ')}`
-      secondaryText = classObject.className
-      caption = `${classObject.tutorInfo.firstName} ${classObject.tutorInfo.lastName}`
-      break;
-    default:
-      break;
-  }
-
-  return (
-    <ChatDisplay
-      chatID={chatID}
-      caption={caption}
-      mainText={mainText}
-      secondaryText={secondaryText}
-      goToChat={goToChat}
-      className={classObject.className}
-      getClasses={getClasses}
-      tutorInfo={classObject.tutorInfo}
-      userInfo={classObject.userInfo}
-      displayNotificationBadge={displayNotificationBadge}
-      clearNotificationCounter={clearNotificationCounter}
-    />
-  )
-}
 
 const ChatPicker: React.FC<IChatPickerProps> = ({ navigation }) => {
   const { loggedUser, setUser, notifications, clearNotificationCounter } = React.useContext(GeneralContext);
