+++ emphasis_education_app/src/components/Chat/common.tsx
@@ -1,10 +1,13 @@
 import * as React from 'react';
+import { View, Text, ActivityIndicator } from 'react-native'
 import {
   CenteredDiv,
   GeneralSpacing,
   ThemedText,
-  FONT_STYLES
+  IconRow,
+  FONT_STYLES,
 } from '../shared';
+import styled from 'styled-components';
 
 const EmptyChatWrapper: React.FC = ({ children }) => (
   <CenteredDiv>
@@ -27,4 +30,46 @@ export const EmptyChat: React.FC = () => (
   <EmptyChatWrapper>
     This is the beginning of your conversation. Let's talk!
   </EmptyChatWrapper>
+);
+
+export const LoadingScreen: React.FC<{ loading: boolean }> = ({ loading }) => (
+  <View>
+    <Text>
+      the chat is deleting
+    </Text>
+    <ActivityIndicator animating={loading} />
+  </View>
+)
+
+export const IconRowLeft = styled(IconRow)`
+  justify-content: flex-start;
+`;
+
+export const LeftText = styled(ThemedText)`
+  padding-right: 5px;
+`
+
+export const RightText = styled(ThemedText)`
+  padding-left: 5px;
+`
+
+export const SpacedItemRow = styled(View)`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+`;
+
+export const NotificationBadge = styled(View)`
+  height: 20px;
+  width: 20px;
+  border-radius: 10px;
+  background-color: grey;
+  margin-right: 10px;
+`
+
+export const ChatContain: React.FC = ({ children }) => (
+  <GeneralSpacing u={10} r={10} d={10} l={10}>
+    {children}
+  </GeneralSpacing>
 );
